user register working without redux integration
