+++ src/actions/types.js
@@ -1,0 +1,3 @@
+export const ADD_USER = 'ADD_USER';
+export const DELETE_USER = 'DELETE_USER';
+export const SELECT_PHOTO = 'SELECT_PHOTO';
