--- src/components/RegisterForm.js
@@ -1,122 +1,0 @@
-import React, { useState } from "react";
-import { View, TextInput, StyleSheet, Text } from "react-native";
-import { Picker } from "@react-native-community/picker";
-import Constants from "expo-constants";
-import { TouchableOpacity } from "react-native-gesture-handler";
-import { connect } from "react-redux";
-import { addUser } from "../actions/user";
-
-function RegisterForm({ addUser, navigation }) {
-
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-  const [accessLevel, setAccessLevel] = useState("");
-
-  return (
-    <View style={styles.container}>
-      <TextInput
-        style={styles.input}
-        importantForAutofill="no"
-        placeholder="First Name"
-        autoCapitalize="none"
-        placeholderTextColor="white"
-        onChangeText={(text) => setFirstName(text)}
-      />
-      <TextInput
-        style={styles.input}
-        importantForAutofill="no"
-        placeholder="Last Name"
-        autoCapitalize="none"
-        placeholderTextColor="white"
-        onChangeText={(text) => setLastName(text)}
-      />
-
-      <View
-        style={{
-          width: 350,
-          height: 55,
-          margin: 20,
-          padding: 8,
-          borderRadius: 14,
-          borderWidth: 1,
-          backgroundColor: "#393e46",
-          justifyContent:"center",
-        }}
-      >
-        <Picker
-          selectedValue={accessLevel}
-          style={{
-
-            color: "white",
-          }}
-          onValueChange={(itemValue, itemIndex) => setAccessLevel(itemValue)}
-        >
-          <Picker.Item label="Select Access Level" value="" />
-          <Picker.Item label="A1" color="black" value="A1" />
-          <Picker.Item label="B2" value="B2" />
-          <Picker.Item label="C3" value="C3" />
-          <Picker.Item label="VIP" value="VIP" />
-        </Picker>
-      </View>
-
-      <TouchableOpacity onPress={() => navigation.navigate("PhotoBooth")}>
-        <Text style={styles.nextScreen}>Take a Profile Picture</Text>
-      </TouchableOpacity>
-      <TouchableOpacity
-        onPress={() =>
-          firstName !== "" && lastName !== "" && accessLevel !== ""
-            ? addUser({
-                firstName: firstName,
-                lastName: lastName,
-                accessLevel: accessLevel,
-              })
-            : alert("Please fill the form and take a picture")
-        }
-      >
-        <Text style={styles.nextScreen}>Add User</Text>
-      </TouchableOpacity>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    marginTop: Constants.statusBarHeight,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "white",
-  },
-  input: {
-    width: 350,
-    height: 55,
-    backgroundColor: "#393e46",
-    margin: 20,
-    padding: 8,
-    color: "white",
-    borderRadius: 14,
-    fontSize: 18,
-    fontWeight: "500",
-    paddingLeft:17
-  },
-  nextScreen: {
-    width: 350,
-    height: 55,
-    backgroundColor: "#00adb5",
-    margin: 20,
-    color: "white",
-    borderRadius: 14,
-    fontSize: 18,
-    textAlign: "center",
-    textAlignVertical: "center",
-    fontWeight: "bold",
-  },
-});
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    newUser: (user) => dispatch(addUser(user)),
-  };
-};
-
-export default connect(null, mapDispatchToProps)(RegisterForm);
