+++ src/components/UserDetails.js
@@ -1,19 +1,31 @@
 import React from "react";
-import { StyleSheet, Text, View } from "react-native";
-import { connect } from "react-redux";
+import { Text, View } from "react-native";
 
-const UserDetails = () => {
+const UserDetails = (props) => {
   return (
-    <View>
-      <Text>UserDetails</Text>
+    <View
+      style={{
+        width: 350,
+        height: 55,
+        backgroundColor: "#393e46",
+        justifyContent: "center",
+        borderRadius: 14,
+        margin: 10,
+      }}
+    >
+      <Text
+        style={{
+          padding: 8,
+          color: "white",
+          fontSize: 18,
+          fontWeight: "500",
+          alignSelf:"center"
+        }}
+      >
+        {props.field}
+      </Text>
     </View>
   );
 };
 
-const styles = StyleSheet.create({});
-
-const mapStateToProps = (state) => ({});
-
-const mapDispatchToProps = {};
-
-export default connect(mapStateToProps, mapDispatchToProps)(UserDetails);
+export default UserDetails;
