+++ src/screens/RegisterScreens/CreateUserScreen.js
@@ -1,27 +1,34 @@
 import React from "react";
-import { StyleSheet, Text, View } from "react-native";
+import { StyleSheet, Text, View, Image } from "react-native";
 import { TouchableOpacity } from "react-native-gesture-handler";
 import Constants from "expo-constants";
+import UserDetails from "../../components/UserDetails";
 
-
-const CreateUserScreen = ({navigation}) => {
+const CreateUserScreen = ({ navigation, route }) => {
   return (
     <View style={styles.container}>
+      <Image
+        source={{ uri: route.params.newUser.userPhoto }}
+        style={{ height: 360, width: 270 }}
+      />
+      <UserDetails field={route.params.newUser.firstName} />
+      <UserDetails field={route.params.newUser.lastName} />
+      <UserDetails field={route.params.newUser.accessLevel} />
+
       <TouchableOpacity
-      onPress={()=>navigation.navigate("RegistrationFormScreen") }
+        onPress={() => navigation.navigate("RegistrationFormScreen")}
       >
-      <Text>RegistrationFormScreen</Text>
+
       </TouchableOpacity>
     </View>
   );
 };
 const styles = StyleSheet.create({
-  container:{
-    flex:1,
-    marginTop:Constants.statusBarHeight,
-    alignItems:"center"
-  }
+  container: {
+    flex: 1,
+    marginTop: Constants.statusBarHeight,
+    alignItems: "center",
+  },
 });
-
 
 export default CreateUserScreen;
